+++ atcg_lib/include/Renderer/RenderPasses/DepthPass.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <Renderer/RenderPass.h>
+#include <Renderer/Texture.h>
+#include <DataStructure/Skybox.h>
+#include <Scene/ComponentRenderer.h>
+
+namespace atcg
+{
+
+
+class DepthPass : public RenderPass
+{
+public:
+    /**
+     * @brief Constructor.
+     */
+    DepthPass(const CullMode cull_mode, const RenderTargetDesc& desc = {});
+
+private:
+};
+}    // namespace atcg